Add files via upload
New version of the LSM9DS1 library for the IMU chip. It was tested on a Nano 33 BLE Sense board and should be backwards compatible with V1.01.0.  

New for all 9 DOF is the possibility:
 to give it calibration zero offset and slope factors. 
 to change the output unit
 to set and get the ODR sample rate frequency ,
 to set and get the internal full scale setting of the chip (IFS) giving it more accuracy at the expense of the range.

Several fixed values in version 1.01 are replaced by functions that produce the value according to the current chip register setting. As a result there are a lot of new set... and get... functions.

Offset and Slope are organised in such way that they can be calibrated separately or together. Their values are independent of the full scale setting or the output unit. Copy their values in a new sketch and it will return calibrated output even when you choose to view it in a different setting. 
The voids  set...Offset  and  set...Slope are made for calibration purposes only. To store their values from read measurements according to the above features. Perhaps they need a better name to reflect this.