400Hz magnetic, DIY calibration, level
setMagnetODR(8); now produces measurements @400Hz. The DIY calibration  is a menu driven program that explains what to do for a simple calibration, does the measurements and produces the data that can be copy/pasted in a sketch. Waterlevel is jus what is says in two directions. The XY compass also produces magnetic inclination.