Runnable on Arduino Uno
The DIY calibration was split into three separate programs. The FS and the ODR settings were added in the menus and in the copy/pasteable output. This increases the accuracy especially for the gyroscope. All the Serial.print("text") 's were encapsulated with the F(..) macro. The IMU.end() now  restores the I2C pullups which brings the power usage back to almost nothing. A small bug in the Full Scale setting of the magnetometer was fixed.